loop/someone/avatar commands, auto-disconnects after 15 min of voice inactivity, some info command UIs cleaned up, clean_everyone added, icon emojis replaced with unicode instead of named-based codes
New commands:
* loop - Loop (or don't loop) a music queue.
* someone - Pings a random guild member with a message.
* avatar - Retrieves a mentioned user's avatar as an attached file.

Bot now auto-disconnects and destroys queue if the voice channel it's in is inactive for 15 min.

Embeds implemented for several info commands.

clean_everyone(ctx, arg) is a lbutil to prevent members without the mention_everyone permission from using the bot to mention @everyone or @here.

Several help icons were replaced with the unicode characters instead of the emojis' names with colons.